Split into components and update GVMI hash
