+++ 02_weather_stats/golem-compose.yml
@@ -23,7 +23,7 @@ payloads:
   app:
     runtime: "vm"
     params:
-      image_hash: "29fd0ccf96f681e4dc8685b2f03aa68ef868f30698d6fe12a089eee9"
+      image_hash: "4a68bb3f791c8db0a8125730fa231941bebcd097f93ecc260d547248"
 nodes:
   cache:
     payload: "cache"
